separeted file for CLI commands.
