+++ platformio.ini
@@ -34,7 +34,7 @@ build_flags =
 lib_deps = 
   bblanchon/ArduinoJson @ 6.20.0
   chrisjoyce911/esp32FOTA @ 0.2.7
-  hpsaturn/ESP32 Wifi CLI @ 0.1.6
+  ; hpsaturn/ESP32 Wifi CLI @ 0.1.7
   https://github.com/256dpi/arduino-mqtt.git#69cb7d2
   https://github.com/tobiasschuerg/InfluxDB-Client-for-Arduino.git#8e5f051
   hpsaturn/CanAirIO Air Quality Sensors Library @ 0.6.4
