+++ src/main.cpp
@@ -257,6 +257,8 @@ void setup() {
     Serial.println("-->[INFO] Sensorslib version\t: " + sensors.getLibraryVersion());
     Serial.println("-->[INFO] enable sensor GPIO\t: " + String(MAIN_HW_EN_PIN));
     logMemory("GPIO");
+    wifiCLIInit();
+    logMemory("CLI");
     startingSensors();
     logMemory("SLIB");
     // Setting callback for remote commands via Bluetooth config
